rename: SelectForm -> SelectInput, apply-celebrity -> celebrity-list로이름 변경
